Match against a div instead of "execution" element (#3)
The latest update has removed the execution element, so we can't use
that as the root of the match when looking for builds any more. There's
a div with the execution class that's in both the old and new UIs
though, so match against that. The logic is a bit funky, cause we want
to match against divs with exactly the "execution" class, and there are
lots of classes that contain execution as a part. So we need to do a
slightly funky match to make sure we only hit exactly the execution
class.